Fix enabling Address Sanitizer on CI

+++ src/engraving/libmscore/score.h
@@ -678,7 +678,7 @@ public:
     void undo(UndoCommand*, EditData* = 0) const;
     void undoRemoveMeasures(Measure*, Measure*, bool preserveTies = false);
     void undoChangeMeasureRepeatCount(Measure* m, int count, staff_idx_t staffIdx);
-    void undoAddBracket(Staff* staff, int level, BracketType type, size_t span);
+    void undoAddBracket(Staff* staff, size_t level, BracketType type, size_t span);
     void undoRemoveBracket(Bracket*);
     void undoInsertTime(const Fraction& tick, const Fraction& len);
     void undoChangeStyleVal(Sid idx, const PropertyValue& v);
